+++ lib/src/dla_gbrfsx_extended.dart
@@ -104,7 +104,7 @@ void dla_gbrfsx_extended(
   // Force HUGEVAL to Inf
   HUGEVAL *= HUGEVAL;
   // Using HUGEVAL may lead to spurious underflows.
-  INCR_THRESH = N.toDouble() * EPS;
+  INCR_THRESH = N * EPS;
   M = KL + KU + 1;
 
   for (J = 1; J <= NRHS; J++) {
