+++ lib/src/dla_syrfsx_extended.dart
@@ -121,7 +121,7 @@ void dla_syrfsx_extended(
   // Force HUGEVAL to Inf
   HUGEVAL *= HUGEVAL;
   // Using HUGEVAL may lead to spurious underflows.
-  INCR_THRESH = N.toDouble() * EPS;
+  INCR_THRESH = N * EPS;
 
   if (lsame(UPLO, 'L')) {
     UPLO2 = ilauplo('L');
